+++ ocrd/ocrd/cli/__init__.py
@@ -31,6 +31,9 @@ from ocrd.cli.resmgr import resmgr_cli
 from ocrd.decorators import ocrd_loglevel
 from .zip import zip_cli
 from .log import log_cli
+from .processing_server import processing_server_cli
+from .processing_worker import processing_worker_cli
+
 
 @click.group()
 @click.version_option()
@@ -48,3 +51,5 @@ cli.add_command(zip_cli)
 cli.add_command(validate_cli)
 cli.add_command(log_cli)
 cli.add_command(resmgr_cli)
+cli.add_command(processing_server_cli)
+cli.add_command(processing_worker_cli)