+++ ocrd/ocrd/decorators/__init__.py
@@ -1,6 +1,8 @@
 from os.path import isfile
 from os import environ
 import sys
+from contextlib import redirect_stdout
+from io import StringIO
 
 import click
 
@@ -10,8 +12,10 @@ from ocrd_utils import (
     set_json_key_value_overrides,
 )
 
-from ocrd_utils import getLogger, initLogging
+from ocrd_utils import getLogger, initLogging, parse_json_string_with_comments
 from ocrd_validators import WorkspaceValidator
+
+from ocrd_network import ProcessingWorker
 
 from ..resolver import Resolver
 from ..processor.base import run_processor
@@ -34,6 +38,8 @@ def ocrd_cli_wrap_processor(
     overwrite=False,
     show_resource=None,
     list_resources=False,
+    queue=None,
+    database=None,
     **kwargs
 ):
     if not sys.argv[1:]:
@@ -50,6 +56,35 @@ def ocrd_cli_wrap_processor(
             list_resources=list_resources
         )
         sys.exit()
+    # If either of these two is provided but not both
+    if bool(queue) != bool(database):
+        raise Exception("Options --queue and --database require each other.")
+    # If both of these are provided - start the processing worker instead of the processor - processorClass
+    if queue and database:
+        initLogging()
+        # TODO: Remove before the release
+        # We are importing the logging here because it's not the ocrd logging but python one
+        import logging
+        logging.getLogger('ocrd.network').setLevel(logging.DEBUG)
+
+        # Get the ocrd_tool dictionary
+        processor = processorClass(workspace=None, dump_json=True)
+        ocrd_tool = processor.ocrd_tool
+
+        try:
+            processing_worker = ProcessingWorker(
+                rabbitmq_addr=queue,
+                mongodb_addr=database,
+                processor_name=ocrd_tool['executable'],
+                ocrd_tool=ocrd_tool,
+                processor_class=processorClass,
+            )
+            # The RMQConsumer is initialized and a connection to the RabbitMQ is performed
+            processing_worker.connect_consumer()
+            # Start consuming from the queue with name `processor_name`
+            processing_worker.start_consuming()
+        except Exception as e:
+            raise Exception("Processing worker has failed with error") from e
     else:
         initLogging()
         LOG = getLogger('ocrd_cli_wrap_processor')